Revert "converted ReportListController and RunReportFormController to use controller annotation"
This reverts commit f4c80fc84ccdf84a00f4cd4753dec03ac79ba4c6.
