+++ omod/src/main/java/org/openmrs/module/reporting/web/reports/RunReportListController.java
@@ -3,23 +3,29 @@
  * v. 2.0. If a copy of the MPL was not distributed with this file, You can
  * obtain one at http://mozilla.org/MPL/2.0/. OpenMRS is also distributed under
  * the terms of the Healthcare Disclaimer located at http://openmrs.org/license.
+ *
  * Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS
  * graphic logo is a trademark of OpenMRS Inc.
  */
 package org.openmrs.module.reporting.web.reports;
 
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
+import javax.servlet.http.HttpServletRequest;
+
+import org.openmrs.api.context.Context;
+import org.openmrs.module.reporting.report.definition.service.ReportDefinitionService;
+import org.springframework.web.servlet.mvc.SimpleFormController;
 
 /**
  *
  */
-@Controller
-public class RunReportListController {
-
-	@RequestMapping("/module/reporting/run/runReport.list")
-	public Object initializeForm() {
-		return "/module/reporting/run/runReportList";
+public class RunReportListController extends SimpleFormController {
+	
+	/**
+	 * @see org.springframework.web.servlet.mvc.AbstractFormController#formBackingObject(javax.servlet.http.HttpServletRequest)
+	 */
+	@Override
+	protected Object formBackingObject(HttpServletRequest request) throws Exception {
+		return Context.getService(ReportDefinitionService.class).getAllDefinitions(false);
 	}
-
+	
 }
